requirements.txt
torch==2.2.0
torchtext==0.17.0
torchvision==0.17.0
tensorboard==2.18.0
numpy==1.26.3
packaging==24.2
pandas==2.2.3
pillow==10.2.0
matplotlib==3.10.0
Flask==3.1.0
ray==2.40.0
ray[train]
scikit-learn==1.6.0
scipy==1.14.1
